move helper to utils and refactor tests
