+++ context/app/static/js/components/detailPage/ProcessedData/ProcessedDataset/hooks.ts
@@ -3,6 +3,7 @@ import { isDataset, type Dataset } from 'js/components/types';
 import { excludeComponentDatasetsClause, getIDsQuery } from 'js/helpers/queries';
 import { useSearchHits } from 'js/hooks/useSearchData';
 import { useProcessedDatasets, type ProcessedDatasetInfo } from 'js/pages/Dataset/hooks';
+import { processDatasetLabel } from 'js/pages/Dataset/utils';
 import { ComponentType } from 'react';
 import { nodeIcons } from '../../DatasetRelationships/nodeTypes';
 
@@ -65,30 +66,6 @@ export function useProcessedDatasetDetails(uuid: string) {
 
   const datasetDetails = searchHits[0]?._source;
   return { datasetDetails, isLoading };
-}
-
-export function processDatasetLabel(
-  dataset: Pick<ProcessedDatasetInfo, 'assay_display_name' | 'pipeline' | 'status' | 'hubmap_id'>,
-  hits: { _source: Pick<ProcessedDatasetInfo, 'assay_display_name' | 'pipeline' | 'status'> }[],
-) {
-  const label = dataset.pipeline ?? dataset.assay_display_name[0];
-  const hasMultipleHitsWithSameLabel =
-    hits.filter((h) => (h._source.pipeline ?? h._source.assay_display_name[0]) === label).length > 1;
-
-  const hasMultipleHitsWithSameLabelAndStatus =
-    hasMultipleHitsWithSameLabel &&
-    hits.filter(
-      (h) => (h._source.pipeline ?? h._source.assay_display_name[0]) === label && h._source.status === dataset.status,
-    ).length > 1;
-
-  if (hasMultipleHitsWithSameLabelAndStatus) {
-    return `${label} (${dataset.status}) [${dataset.hubmap_id}]`;
-  }
-  if (hasMultipleHitsWithSameLabel) {
-    return `${label} (${dataset.status})`;
-  }
-
-  return label;
 }
 
 export function useProcessedDatasetTabs(): { label: string; uuid: string; icon: ComponentType | undefined }[] {
